refactor: props/state 타입 위치 컴포넌트로 이동
