+++ src/components/FilterSection/index.tsx
@@ -1,7 +1,6 @@
 import './style.css';
 import { Component } from 'react';
 import { FilterOption } from '../../types';
-import { FilterSectionProps, FilterSectionState } from '../../types/component';
 import {
   DEFAULT_CATEGORY,
   DEFAULT_SORT_BY,
@@ -10,6 +9,15 @@ import {
   SORT_BY,
 } from '../../constants';
 import Select from '../Select';
+
+interface FilterSectionProps {
+  onChange: CallableFunction;
+}
+
+interface FilterSectionState {
+  category: string;
+  sortBy: string;
+}
 
 class FilterSection extends Component<FilterSectionProps, FilterSectionState> {
   state: FilterSectionState;
